Load `deck2019.js` script with `defer`
`script` tags loaded with the `defer` attribute download in parallel to
the DOM being built but are executed after the DOM is fully constructed.

This change speeds up the page by letting the script be downloaded in
parallel and allows removing the `DOMContentLoaded` event listener for
setting up Reveal; instead the code can just be executed directly by the
loaded JavaScript module.